Organize jobs into "repeating" folder
